# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Wishaw YMCA Badge API
  version: 0.0.1
  description: |
    OpenAPI contract for the Wishaw YMCA Esports badging backend.

    This API is designed for a separate React frontend and uses session-based authentication.
    Successful login establishes a JSESSIONID cookie; subsequent calls must include that cookie.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
  /api/v1/auth/login:
    post:
      tags:
        - Auth
      summary: Log in and create a backend session
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: superadmin
              password: admin123
      responses:
        '200':
          description: Login succeeded and a session cookie is returned.
          headers:
            Set-Cookie:
              description: Session cookie carrying the authenticated Spring session.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/auth/logout:
    post:
      tags:
        - Auth
      summary: End the current backend session
      operationId: logout
      security: []
      responses:
        '204':
          description: Session invalidated.
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/auth/me:
    get:
      tags:
        - Auth
      summary: Get the currently authenticated user
      operationId: getAuthenticatedUser
      security:
        - SessionCookieAuth: []
      responses:
        '200':
          description: Authenticated user summary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/admin/users:
    post:
      tags:
        - Admin Users
      summary: Create a user account
      operationId: createUser
      security:
        - SessionCookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummaryResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'
    get:
      tags:
        - Admin Users
      summary: List users
      operationId: listUsers
      security:
        - SessionCookieAuth: []
      responses:
        '200':
          description: All users visible to the admin scope.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSummaryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/admin/users/{id}:
    get:
      tags:
        - Admin Users
      summary: Get a user by id
      operationId: getUser
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummaryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      tags:
        - Admin Users
      summary: Update a user account
      operationId: updateUser
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Updated user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummaryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/admin/users/{id}/status:
    patch:
      tags:
        - Admin Users
      summary: Activate or deactivate a user
      operationId: updateUserStatus
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateActiveStatusRequest'
      responses:
        '200':
          description: Updated user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummaryResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/admin/parents/link-player:
    post:
      tags:
        - Admin Users
      summary: Link a parent account to a player account
      operationId: linkParentToPlayer
      security:
        - SessionCookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkParentRequest'
      responses:
        '204':
          description: Parent link created.
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/admin/centres:
    post:
      tags:
        - Centres
      summary: Create a centre
      operationId: createCentre
      security:
        - SessionCookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCentreRequest'
      responses:
        '201':
          description: Centre created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CentreResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'
    get:
      tags:
        - Centres
      summary: List centres
      operationId: listCentres
      security:
        - SessionCookieAuth: []
      responses:
        '200':
          description: Centre list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CentreResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/admin/centres/{id}:
    put:
      tags:
        - Centres
      summary: Update a centre
      operationId: updateCentre
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CentreId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCentreRequest'
      responses:
        '200':
          description: Updated centre.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CentreResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/v1/admin/enrollments:
    post:
      tags:
        - Enrollments
      summary: Enroll a player on a module
      operationId: createEnrollment
      security:
        - SessionCookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEnrollmentRequest'
      responses:
        '201':
          description: Enrollment created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/admin/enrollments/{id}/status:
    patch:
      tags:
        - Enrollments
      summary: Update enrollment status
      operationId: updateEnrollmentStatus
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/EnrollmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEnrollmentStatusRequest'
      responses:
        '200':
          description: Updated enrollment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/v1/leaderboards/centre/{centreId}:
    get:
      tags:
        - Leaderboards
      summary: Get the leaderboard for one centre
      operationId: getCentreLeaderboard
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CentreId'
      responses:
        '200':
          description: Centre leaderboard.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaderboardEntryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    SessionCookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Session cookie returned by POST /api/v1/auth/login.
  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    CentreId:
      name: centreId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    EnrollmentId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    BadRequest:
      description: Request was syntactically valid but semantically invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    ValidationError:
      description: Bean validation failed for one or more request fields.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Authentication is missing, invalid, or expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Forbidden:
      description: Authenticated user does not have the required role.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Conflict:
      description: Request conflicts with current state, usually a duplicate resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    ServerError:
      description: An unexpected backend error occurred.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    CreateUserRequest:
      type: object
      required:
        - username
        - password
        - displayName
        - role
      properties:
        username:
          type: string
        password:
          type: string
          format: password
        displayName:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        centreId:
          type: integer
          format: int64
          nullable: true
        groupId:
          type: integer
          format: int64
          nullable: true
        externalRef:
          type: string
          nullable: true
    UpdateUserRequest:
      type: object
      properties:
        displayName:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        centreId:
          type: integer
          format: int64
          nullable: true
        groupId:
          type: integer
          format: int64
          nullable: true
        externalRef:
          type: string
          nullable: true
    UpdateActiveStatusRequest:
      type: object
      required:
        - active
      properties:
        active:
          type: boolean
    LinkParentRequest:
      type: object
      required:
        - parentUserId
        - playerUserId
      properties:
        parentUserId:
          type: integer
          format: int64
        playerUserId:
          type: integer
          format: int64
        relationshipLabel:
          type: string
          nullable: true
    CreateCentreRequest:
      type: object
      required:
        - name
        - code
      properties:
        name:
          type: string
        code:
          type: string
...
    CreateEnrollmentRequest:
      type: object
      required:
        - playerId
        - moduleId
      properties:
        playerId:
          type: integer
          format: int64
        moduleId:
          type: integer
          format: int64
        groupId:
          type: integer
          format: int64
          nullable: true
    UpdateEnrollmentStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/EnrollmentStatus'
...
    AuthResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
        displayName:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        message:
          type: string
    UserSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        displayName:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        active:
          type: boolean
        centreId:
          type: integer
          format: int64
          nullable: true
        centreName:
          type: string
          nullable: true
        groupId:
          type: integer
          format: int64
          nullable: true
        groupName:
          type: string
          nullable: true
        externalRef:
          type: string
          nullable: true
    CentreResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        code:
          type: string
        active:
          type: boolean
...
    EnrollmentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        playerId:
          type: integer
          format: int64
        playerName:
          type: string
        moduleId:
          type: integer
          format: int64
        moduleName:
          type: string
        status:
          $ref: '#/components/schemas/EnrollmentStatus'
...
    LeaderboardEntryResponse:
      type: object
      properties:
        rank:
          type: integer
        playerId:
          type: integer
          format: int64
        displayName:
          type: string
        centreName:
          type: string
          nullable: true
        groupName:
          type: string
          nullable: true
        totalPoints:
          type: integer
        highestLevel:
          type: string
          nullable: true
...
    ApiErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
    Role:
      type: string
      enum:
        - SUPER_ADMIN
        - CENTRE_ADMIN
        - PLAYER
        - PARENT
    EnrollmentStatus:
      type: string
      enum:
        - ASSIGNED
        - IN_PROGRESS
        - COMPLETED
        - ARCHIVED
...